Refactor referral search api
