+++ client/hooks/api/referral/useSearchReferral.tsx
@@ -1,0 +1,43 @@
+import apiService from "@/utils/common/api"
+import { supabase } from "@/utils/services/supabase/config"
+import { useInfiniteQuery } from "@tanstack/react-query"
+
+import { QueryKeyString } from "@/types/common/query-key-string"
+import { ReferralType } from "@/types/common/referral-type"
+
+interface ISearchRefererFilterMeta {
+  companyName: string
+  cityUuid?: string
+  provinceUuid?: string
+  countryUuid?: string
+  industryUuid?: string
+  sorting: string
+  yoeMax?: string
+  yoeMin?: string
+}
+
+const useSearchReferral = (
+  sorting: string,
+  filterMeta: ISearchRefererFilterMeta,
+  type: ReferralType
+) => {
+  const keyString =
+    type === ReferralType.REFEREE
+      ? QueryKeyString.SEARCH_REFEREE
+      : QueryKeyString.SEARCH_REFERRER
+  return useInfiniteQuery({
+    queryKey: [keyString, { sorting, filterMeta, type }],
+    queryFn: apiService.searchReferral,
+    refetchOnMount: false,
+    refetchOnWindowFocus: false,
+    getNextPageParam: (lastPage, allPages: any[]) => {
+      if (lastPage && lastPage.length > 0) {
+        return allPages.length
+      } else {
+        return null
+      }
+    },
+  })
+}
+
+export default useSearchReferral
