+++ client/utils/common/api/index.tsx
@@ -87,6 +87,92 @@ const apiService = {
       throw err
     }
   },
+  searchReferral: async ({ pageParam = 0, queryKey }: any) => {
+    try {
+      const NUMBER_OF_DATE_PER_FETCH = 10
+      const countryUuid = queryKey[1].filterMeta.countryUuid
+      const provinceUuid = queryKey[1].filterMeta.provinceUuid
+      const cityUuid = queryKey[1].filterMeta.cityUuid
+      const industryUuid = queryKey[1].filterMeta.industryUuid
+      const companyName = queryKey[1].filterMeta.companyName
+      const yoeMax = queryKey[1].filterMeta.yoeMax
+      const yoeMin = queryKey[1].filterMeta.yoeMin
+      const type = queryKey[1].type satisfies ReferralType
+
+      const sort = queryKey[1].sorting.split(",")
+      const order = sort[1] === "dec" ? false : true
+      const from = pageParam * NUMBER_OF_DATE_PER_FETCH
+      const to = from + NUMBER_OF_DATE_PER_FETCH
+
+      let query = supabase
+        .from("user")
+        .select(
+          `
+            uuid,
+            email,
+            username,
+            avatar_url,
+            description,
+            company_name,
+            job_title,
+            year_of_experience,
+            social_media_url,
+            country(
+              cantonese_name
+            ),
+            province(
+              cantonese_name
+            ),
+            city(
+              cantonese_name
+            ),
+            industry(
+              cantonese_name
+            ),
+            is_referer,
+            is_referee
+          `
+        )
+        .lte("year_of_experience", yoeMax ? parseInt(yoeMax) : 100)
+        .gte("year_of_experience", yoeMin ? parseInt(yoeMin) : 0)
+        .order("year_of_experience", { ascending: order })
+        .order("id", { ascending: true })
+        .range(from, to)
+
+      if (type === ReferralType.REFERRER) {
+        query = query.eq("is_referer", true)
+      }
+
+      if (type === ReferralType.REFEREE) {
+        query = query.eq("is_referee", true)
+      }
+
+      if (countryUuid) {
+        query = query.eq("country_uuid", countryUuid)
+      }
+      if (provinceUuid) {
+        query = query.eq("province_uuid", provinceUuid)
+      }
+      if (cityUuid) {
+        query = query.eq("city_uuid", cityUuid)
+      }
+      if (industryUuid) {
+        query = query.eq("industry_uuid", industryUuid)
+      }
+
+      if (companyName.length > 0) {
+        query = query.ilike("company_name", `%${companyName}%`)
+      }
+
+      const { data, error } = await query
+
+      if (error) throw error
+
+      return data
+    } catch (err) {
+      throw err
+    }
+  },
 
   // Post
   createPost: async (req: ICreatePostRequest) => {
@@ -178,7 +264,9 @@ const apiService = {
       if (industryUuid !== undefined) {
         query = query.eq("industry_uuid", industryUuid)
       }
-      const { data } = await query.order("id", { ascending: true })
+      const { data, error } = await query.order("id", { ascending: true })
+
+      if (error) throw error
 
       return data
     } catch (err) {
